朝向、移动方向、激活状态、装载、旋转方向、旋转起始位置、延时触发、触发周期

+++ GameAClient/Assets/Game/Scripts/Battle/Skill/Bullet/Bomb.cs
@@ -134,7 +134,7 @@ namespace GameA.Game
             }
         }
 
-        protected override void UpdateAngle(int angle, EDirectionType eDirectionType)
+        protected override void UpdateAngle(float angle, EDirectionType eDirectionType)
         {
             _angle = angle;
             var rad = _angle * Mathf.Deg2Rad;
